refactor(desktop): extract generate_media configuration into a focused module
extensions-view.tsx had grown to 2,110 lines after absorbing the media
types, selection validation, and configuration UI. Move MediaTypeConfiguration,
GenerateMediaToolConfig, isValidMediaSelection, MediaModelConfiguration,
GenerateMediaConfiguration, and the expandable generate_media tool card into
generate-media-tool.tsx, leaving extensions-view.tsx with the generic tool
list (1,840 lines, back near its pre-feature size).

The card now owns its expansion state, so the view no longer tracks
expandedToolIds. Shared provider fixtures move to a test-fixtures module;
generate-media-tool.test.tsx adds direct contract tests for selection
validity, setup gating, catalog-refresh gating, and the no-config fallback,
while extensions-view.test.tsx keeps the integration tests through
CustomizationSectionView.
